Medical assistant/Whole heart segmentation/requirements.txt
numpy~=1.20
pandas~=1.1.5
matplotlib~=3.3.4
scipy~=1.6.1
nibabel~=3.2.1
opencv-python~=4.5.1.48
Pillow~=8.1.2
tensorflow~=2.4.1
tensorflow-cloud~=0.1.14
itk~=5.1.2
nipype~=1.6.0
h5py~=2.10.0
scikit-image~=0.17.2